Detects object and causes shadow to render
Object is detected using two raycasts. If both are detecting the object only then the shadow is being cast. the animation completion is detected using signal and after rise animation is complete default is played. The shadow is part of the object only. It gets displayed when light falls on it. It is not instantiated from code.