Added sound to walking, fixed collision box bug
added sound to walking.

fixed collision box bug : even when shadow was not appearing player was not able to movev through that place because collision box was enabled. so collision box has been disabled. gravity on this body has been disabled too because the body falls through as soon as collision box is disabled even if the sprite gets hidden.