Completed mushroom collection code
CollectibleObjectArea has a script attached to it that calls the parents get_collected_by_player when player enters its area2D. This node is a child to each of the mushroom types. Each mushroom has a script attached that has a method get_collected_by_player which increases the count of item category in playerCollectibleManager and then removes the node from the tree. The logic updating the display of the fillled slots has also been cleaned and now the UI will show the exact number of mushrooms collected.